verify_key.py
import google.generativeai as genai
import sys

api_key = sys.argv[1]

try:
    genai.configure(api_key=api_key)
    print("Listing available models:")
    for m in genai.list_models():
        if 'generateContent' in m.supported_generation_methods:
            print(f"- {m.name}")
except Exception as e:
    print(f"Error listing models: {e}")
    sys.exit(1)
