pyproject.toml
[project]
name = "gemini-chat-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "google-generativeai>=0.8.5",
    "uvicorn>=0.38.0",
]

[[tool.uv.index]]
url = "https://pkgs.safetycli.com/repository/public/pypi/simple/"
default = true
